# api/api.yml
openapi: 3.1.0
info:
  title: OpenAPI
  description: Notes API for Learning
  version: 1.0.0
...
servers:
  - url: http://localhost:9999/api/rest
...
paths:
  /notes:
    post:
      tags:
        - note
      summary: Create new Note
      operationId: createNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note' 
        required: true 
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
    get:
      tags:
        - note
      summary: Get All Notes with limit and offset
      operationId: getAllNotes
      parameters:
        - name: limit
          in: query
          description: Limit
          required: false
          schema:
            type: integer
            format: int64
            default: 10
        - name: offset
          in: query
          description: offset
          required: false
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
  /notes/{noteId}:
    get:
      tags:
        - note
      summary: Find note by ID
      description: Returns a single Note
      operationId: getNoteById
      parameters:
        - name: noteId
          in: path
          description: ID of Note to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Note:
      required:
        - message
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
        tags:
          type: array
          items:
            type: string
    Error:
      required:
        - error
      type: object
      properties:
        error:
          type: string